CV-Reformatter-POC/_archive-20251009-130955/requirements.txt
python-docx>=1.1.2
pytest>=8.0.0

mammoth>=1.6.0
